requirements.txt
Jinja2==3.1.6
PyYAML==6.0.2
pytest==8.3.5
